Fixes #24
I’m not sure why we’re dispatching both a “selctedDateChanged” event and a “changed” event for the same action. It seems like we should dispatch one or the other.
On the one hand, “change” is a standard name, so it’s easily discoverable. On the other hand, “change” is one of the special events which become BrowserEvents when dispatched.
I think the change event should be removed.