Update AMFBinaryData.as
Add new variable "flexTraitFound" and change readScriptObject() to allow ArrayCollection and ObjectProxy objects sent from CF to be correctly decoded. This change should not affect anything other than this. (Note that ObjectProxy is not tested, so a further change could be necessary and I will do this when I can test it with an ObjectProxy.)