Frameworks: disable JS tests in main build for now
On macOS, the RoyaleUnit task keeps control over the WebSocket's port for a short time after the task has completed. When this happens, running the RoyaleUnit task again fails because the port is still not available. Can't enable JS tests in main build until this is fixed. Doesn't seem to affect Windows.