Maven Distribution: does not require Adobe artifacts to create a distribution
If distributionTargetFolder is specified, Adobe artifacts are still required. The change is that a zip/tar.gz distribution does not include Adobe dependencies because they were not actually used. We could potentially use the Apache playerglobal for distributionTargetFolder, but we don't have a full alternative to the AIR SDK (only to airglobal).