Fix: Scroll in VirtualDataGrid was unusable with more than 2 columns
VirtualDataGrid was not usable for more than 2 columns because as soon the user made a faster vertical scroll or reposition to another line, all the VirtualDataGrid was messed up.
With this fix (tested with 6 columns and thousands of records) this fix finally fix this issue.
Also, currently to have the VirtualDataGrid working propertly with many records, the DataGridScrollSpeed bead with a minimum speed about 0.5 is needed to emulate the scroll at same speed.