MXMLDataInterpreter: fixed styles/effects/events not getting handled for objects that don't implement IStrand
In particular, this fixes <mx:method> for RemoteObject. It's supposed to generate an Operation which has result/fault events, and the listeners were being ignored instead of getting added as intended. This would have been broken for any class that is an event dispatcher, but not a strand, though, so it's certainly an improvement in general, and not just for this particular case.

A future update might consider refactoring generateMXMLObject() and initializeStrandBasedObject() to consolidate duplicate code and branch only for the strand-specific parts. This should put the output file size back to around the original size (and, I suspect, actually smaller than the original size!)