Fix case where sortable=false
This change fixes the problem where sortable is still possible even though sortable=false is set in the mxml.

Please note that this does not fix the fact that all columns are sortable by default while they shouldn't. Haven't found a fix for that yet.

Let me know if the fix seems correct to you.