Adding DOAP file draft
I have created a DOAP file, per a discussion on board@a.o that reminded me that we do not seem to have one for Royale. I did not add information about an 'implemented standard' because I do not know what standard applies to Royale.

I invite suggestions to improve this information. Once this file is in good-enough shape, I will register it per the instructions at https://projects.apache.org/create.html. I will undertake to update it as necessary (mainly when we release a new version).