MXRoyale: fix BitmapAsset constructor first parameter type
super() constructor accepts BitmapData on SWF and Object on JS. Parameter was incorrectly typed as Bitmap, which isn't correct for either one. Now has separate constructors for SWF and JS that accept the correct types.

BitmapAsset in Flex is related to [Embed] metadata, so I suspect that this class may not even be used in Royale normally.