Network: add XMLResult and JSONResult beads for HTTPService
When the HTTPService dispatches its complete event, the bead parses the text in HTTPService.data as either XML or JSON.

HTTPService already has a json getter that parses the response text as JSON. However, it doesn't have one for XML, and we may want to support additional text-based formats in the future (such as ini, yaml, toml, or jsonc... off the top of my head), so this would be the PAYG way to add new formats. JSONResult is redundant with the json getter, but it could potentially be used by other new service classes, so it's probably good to have for completeness.